+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - chore: bumps .nvmrc and nodejs version in CI to 22
+- chore: npm audit fixes
 
 ## [2.2.0] - 2024-12-12