+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Changed
 
+- feat: HttpAgent uses anonymous identity to make syncTime call, which can allow readState calls to work beyond 5 minutes
 - chore: bumps .nvmrc and nodejs version in CI to 22
 - HttpAgent now awaits fetching rootkey before making network calls if `shouldFetchRootKey` is set
 - chore: npm audit fixes